feat(mpu6050, main): 启用MPU6050并优化调试配置
- 为mpu6050.cpp的Wire.requestFrom添加显式uint8_t类型转换以消除编译警告
- 取消MPU6050初始化与校准的注释，启用传感器功能
- 调整调试宏默认值：关闭PID打印，开启串口调试打印
- 添加IMU角度数据的串口调试打印逻辑
- 补充main.cpp中的代码注释并调整空白格式
- 修复mpu6050.cpp文件末尾缺少换行的问题

+++ ESP32S3/helloWorld/lib/mpu6050/src/mpu6050.cpp
@@ -40,7 +40,7 @@ bool mpu6050_init(uint8_t addr)
     Wire.beginTransmission(_mpu_addr);
     Wire.write(0x75);
     Wire.endTransmission();
-    Wire.requestFrom(_mpu_addr, 1);
+    Wire.requestFrom(static_cast<uint8_t>(_mpu_addr), static_cast<uint8_t>(1));
     uint8_t id = Wire.read();
     if(id != 0x70)
     {
@@ -73,7 +73,7 @@ static void mpu6050_read_raw(void)
     Wire.beginTransmission(_mpu_addr);
     Wire.write(0x3B);
     Wire.endTransmission(false);
-    Wire.requestFrom(_mpu_addr, 14);
+    Wire.requestFrom(static_cast<uint8_t>(_mpu_addr), static_cast<uint8_t>(14));
 
     ax = Wire.read() << 8 | Wire.read();
     ay = Wire.read() << 8 | Wire.read();
